third_party/WebKit/LayoutTests/accessibility/table-headers-expected.txt

No	Month	Expenses
1	Jan	100
2	Feb	150
3	Mar	200
Check whether column and row headers are reported correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.rowHeaderAtIndex(0).isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS table.rowHeaderAtIndex(1).isEqual(table.cellForColumnAndRow(0, 1)) is true
PASS table.rowHeaderAtIndex(2).isEqual(table.cellForColumnAndRow(0, 2)) is true
PASS table.rowHeaderAtIndex(3).isEqual(table.cellForColumnAndRow(0, 3)) is true
PASS table.columnHeaderAtIndex(0).isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS table.columnHeaderAtIndex(1).isEqual(table.cellForColumnAndRow(1, 0)) is true
PASS table.columnHeaderAtIndex(2).isEqual(table.cellForColumnAndRow(2, 0)) is true
PASS successfullyParsed is true

TEST COMPLETE
